pyproject.toml
[project]
name = "marv-ai-agent"
version = "0.1.0"
description = "A CLI AI agent to create Google Ads campaigns using LangChain, MCP, and MySQL"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain",
    "openai",
    "requests",
    "python-dotenv",
    "rich",
    "sqlalchemy",
    "pymysql"
]
